src/PhotoSearch.Florence2.API/src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "syamaner", email = "syamaner@users.noreply.github.com" }
]
dependencies = [
    "einops>=0.8.0",
    "flask>=3.0.3",
    "opentelemetry-distro[otlp]>=0.45b0",
    "opentelemetry-instrumentation-flask>=0.45b0",
    "pillow>=10.2.0",
    "timm>=1.0.7",
    "transformers[torch]>=4.42.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/src"]
